+++ src/modules/ingest/ingest.module.ts
@@ -1,9 +1,13 @@
 import { Module } from "@nestjs/common";
+import { IngestController } from "./ingest.controller";
+import { IngestService } from "./ingest.service";
+import { StorageModule } from "../storage/storage.module";
+import { AuthModule } from "../auth/auth.module";
 
 @Module({
-  imports: [],
-  controllers: [],
-  providers: [],
-  exports: [],
+  imports: [AuthModule, StorageModule],
+  controllers: [IngestController],
+  providers: [IngestService],
+  exports: [IngestService],
 })
 export class IngestModule {}